+++ packages/spacecat-shared-utils/src/s3.js
@@ -10,8 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
-import AWSXray from 'aws-xray-sdk';
 import { S3Client } from '@aws-sdk/client-s3';
+import { instrumentAWSClient } from './xray.js';
 
 /**
  * Adds an S3Client instance and bucket to the context.
@@ -29,7 +29,7 @@ export function s3Wrapper(fn) {
         S3_BUCKET_NAME: bucket,
       } = context.env;
 
-      context.s3.s3Client = AWSXray.captureAWSv3Client(new S3Client({ region }));
+      context.s3.s3Client = instrumentAWSClient(new S3Client({ region }));
       context.s3.s3Bucket = bucket;
     }